backend/pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=0.1.0",
    "drf-social-oauth2>=3.1.0",
    "psycopg2>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.1.0",
]
